Implemented support for GitLab cloud. Updated gitlab4j version.

+++ src/main/java/com/microfocus/octane/gitlab/helpers/PullRequestHelper.java
@@ -44,7 +44,7 @@ public class PullRequestHelper {
                                                 SCMRepository targetScmRepository,
                                                 List<SCMCommit> mergeRequestCommits) {
         return DTOFactory.getInstance().newDTO(PullRequest.class)
-                .setId(Integer.toString(mergeRequest.getIid()))
+                .setId(Long.toString(mergeRequest.getIid()))
                 .setTitle(mergeRequest.getTitle())
                 .setDescription(mergeRequest.getDescription())
                 .setState(mergeRequest.getState())
